+++ ModuleSceneSpace.h
@@ -19,9 +19,9 @@ public:
 	//SDL_Texture* stars;
 
 	//ROF ----
-	int scroll_camera_speed;
-	int scroll_player_speed;
-	int limit_xneg, limit_xpos;
+	float scroll_camera_speed;
+	float scroll_player_speed;
+	float limit_xneg, limit_xpos;
 
 	ModuleSceneSpace(Application *app, bool start_enabled = true);
 	~ModuleSceneSpace();
